feat(Modal): onOpened and onClosed callbacks for modals (#434)
add onOpened and onClosed callbacks for modals
fix logic for where new props are called. re-label methods to align with purpose

Closes #306

Breaks Modal onEnter and onExit props are not longer based on when the modal is opened and closed, they are based on when the component mounts and unmounts. onOpened and onClosed are now used for when the modal opens and closes.